requirements.txt
reportlab
pmdarima
tensorflow
openpyxl
numpy
pandas
matplotlib
seaborn
scikit-learn
statsmodels
Pillow
